last

+++ js/home.js
@@ -1,8 +1,9 @@
 document.addEventListener("DOMContentLoaded", () => {
-  let loginState = sessionStorage.getItem('IsLogin');
-  let LoginUserName = sessionStorage.getItem('LoginUserName');
+  let loginState = sessionStorage.getItem("IsLogin");
+  let LoginUserName = sessionStorage.getItem("LoginUserName");
   let signBTN = document.getElementById("userlink");
   let NavUsername = document.getElementById("userName");
+  let acessoriesBTN = document.getElementById("acessories");
 
   function xloading() {
     console.log("xloading function called");
@@ -10,19 +11,31 @@ document.addEventListener("DOMContentLoaded", () => {
     console.log("LoginUserName:", LoginUserName);
 
     if (Boolean(loginState)) {
-      signBTN.innerHTML = 'SignOut';
-      NavUsername.innerHTML = ('Welcome .... |'+LoginUserName);
+      signBTN.innerHTML = "SignOut";
+      NavUsername.innerHTML = "Welcome .... " + LoginUserName;
     } else {
-      signBTN.innerHTML = 'SignIn';
-      NavUsername.innerHTML = '';
+      signBTN.innerHTML = "SignIn";
+      NavUsername.innerHTML = "";
     }
   }
-  signBTN.addEventListener("click",function(){sessionStorage.setItem("IsLogin", false);
-      loginState = false;
-      sessionStorage.clear();
-          });
+  signBTN.addEventListener("click", function () {
+    sessionStorage.setItem("IsLogin", false);
+    loginState = false;
+    sessionStorage.clear();
+  });
 
-  if (signBTN && NavUsername) { // check if elements are found
+  acessoriesBTN.addEventListener("click", function () {
+    if (Boolean(loginState)) {
+      
+        location.assign("../product.html");
+     }else{
+        location.assign("../login.html");
+        sessionStorage.clear();
+    }
+  });
+
+  if (signBTN && NavUsername) {
+    // check if elements are found
     window.addEventListener("load", xloading);
   } else {
     console.log("Error: HTML element not found");
